revert changes to add non-local names to internal map
Merge Request !1489 made a change to which names were stored in the internal
map. Following this, non-local names (user-provided ones; not starting with '%')
were stored in the internal map as well as local names. This inadvertently broke
some fdp and circo assumptions (#1876, #1877, !1676).

This change reverts the main pieces of the following commits, resolving #1876
and #1877, while re-opening #1767 and #1789.
  * 4f283dd1c02a6a4999b53ad2fcbf2264a7074a8b
  * 85b09cf13179b0e5ab8bddb4857e3d2af0a39a31
  * 9409324489a69557229d3d6f505857b9af85a913
  * 2a9449a99b2a2146fce01fa1d9713e999ad3dd4e
  * 14be5169ef49faad0f30cd9d36cdd438e1739f77
  * b6ffeca3a4457efcffbc3fcdbcee683375f74d05

We will need to find a different solution to #1767.