GVC: export `gvToolTred`
This is a temporary bandaid solution for a missing export from the GVC
library.  The general solution of fixing the annotation on the
declarations is better suited for Windows.  That would allow the
compiler to make appropriate optimizations.  Modern MinGW should also be
able to handle the exports through the attribute, avoiding the need to
maintain the extra def file.

The attributes could also be further generalized to aid on Unix
platforms.

This partially addresses #2507.