remove some unhelpful '-L' suppression in Autotools build
These paths were suppressed in d4e2eec15daadd1bca60d20ffce022bea845621d
presumably because they are assumed to be in the compiler’s sysroot and are
therefore unnecessary. This is incorrect. These paths are not guaranteed to
be in the compiler’s known library directories or even to be known to the
compiler at all.