CMake: hard code GD features on Ubuntu/Debian/etc
The Debian packaging ecosystem ships a libgd with neither `gdlib-config` nor
a PkgConfig (PC) file. The result is that it is not possible to do GD feature
detection. To work around this, hard code the features we believe to be
available. This change mirrors logic from the Autotools build system.

Gitlab: #1786