tests: port Criterion test cases to standalone C
In the past, the Criterion framework was used to unit test Graphviz. This part
of the test suite has long been disabled, with Criterion unused.

This change converts each existing Criterion test case to a standalone C program
and adds a Pytest test case wrapper for invoking them. The intention is to move
away from a dependency on Criterion.

Note that the test cases have somewhat rotted and will need to be updated before
they can be re-enabled.