cgraph: rename 'stack_push' to 'stack_try_push'
Now that the generic stack implementation is widely rolled out, it is apparent
that few callers want to handle push errors. I.e. most call
`stack_push_or_exit`. So it seems to make more sense to rename this to the more
readable `stack_push`. This commit prepares for this by renaming the existing
`stack_push` to something more explicit.