cgraph: make 'FIRST_ALLOCATION' a const instead of enum
This improves the ability of this code to be included by a C++ source. Without
this, the compiler says:

  error: enumerated and non-enumerated type in conditional expression
    [-Werror=extra]

Apparently the C++ compiler considers the result types of a ternary expression
to be strongly typed in a way that C does not. This code is still not quite
includable in C++ (the return of `realloc` needs an explicit cast), but this
gets us closer to this goal.

Two questions may occur to the reader:

  1. Why would you include this header into a C++ source? Just use
     `std::vector`.

  2. Why was this an enum in the first place?

In answer to (1), you would not deliberately, but this happens transitively in
an upcoming commit. list.h ends up used in a header that is included by both C
and C++ sources, resulting in the C++ compiler seeing this header.

In answer to (2), because it is what I picked when writing the original. This is
subjective, but my opinion on your options to define a numeric constant in C:

  ┌──────────────┬───────────────────────┬───────────────────────┬──────────────────┐
  │ option       │ usable as array size? │ debug symbol for GDB? │ constant folded? │
  ╞══════════════╪═══════════════════════╪═══════════════════════╪══════════════════╡
  │ #define      │ yes                   │ no                    │ yes              │
  │ enum         │ yes, except for MSVC  │ yes                   │ yes              │
  │ static const │ no                    │ yes                   │ maybe            │
  └──────────────┴───────────────────────┴───────────────────────┴──────────────────┘

From this table, enum provides the best alternative where you get a debug symbol
and guaranteed constant folding. IMHO.