xdot: when printing a truncated double, avoid transitioning through an int
The prior code printed a double without its digits past the decimal point by
calling `printInt`. The bounds of the int type (`INT_MIN`, `INT_MAX`) are not
guaranteed to exceed the bounds of the double type (`-DBL_MAX`, `DBL_MAX`), and
in practice they are usually much smaller. As a result of this, printing a
double this way could overflow leading to an unexpected output. We can avoid
this issue by doing all intermediate work exclusively on doubles. This change
squashes 6 -Wfloat-conversion warnings.