gvpr: fix references to non-volatile variables after 'longjmp' return
When using `longjmp` to return to a previous location where `setjmp` was called,
all variables of automatic storage have indeterminate values. In practice this
means all your non-volatile local variables and function parameters.

We have been slowly unwinding uses of `longjmp` in the code base due to the
complexity of its side conditions like this. However in this instance it is not
straightforward to remove it because it is part of the user-exposed behavior.

To deal with it we instead mark anything used after the `longjmp` returns as
volatile.

Gitlab: #1801