cgraph: prototype 'aagerror', squashing a -Wmissing-prototypes warning
Normally Flex prototypes the `yyerror` function for you. But we change the
default prefix of this scanner from `yy` to `aag`. So this function ends up
unprototyped and the compiler thinks this worthy of a warning.

In a sense, this change is a bit of a hack to silence this. But there does not
seem much value to prototyping this in a separate header because most of the
calls to this function come from generated Flex code itself.