API BREAK: common: use cgraph wrapper for allocation in 'elist_append'
The lib/cgraph/alloc.h wrappers are similar to the older lib/common/memory.h
wrappers except (1) they are header-only and (2) they live in a directory
(cgraph) that is at the root of the dependency tree. The long term plan is to
replace all use of lib/common/memory.h with lib/cgraph/alloc.h.

It looks a bit strange putting a reference to an internal function,
`gv_recalloc`, that is not even in scope into a public header. But this is
essentially equivalent to the previous code. `ALLOC` is also internal and is not
in scope here. Among other concerns, this is pretty strong evidence that no
users are relying on this macro.