common init_graph: use cgraph wrappers for allocation
The lib/cgraph/alloc.h wrappers are similar to the older lib/common/memory.h
wrappers except (1) they are header-only and (2) they live in a directory
(cgraph) that is at the root of the dependency tree. The long term plan is to
replace all use of lib/common/memory.h with lib/cgraph/alloc.h.

Note that due to the change in the previous commit, these lists are known NULL
with sizes zeroed at this point. So there is no need to use the `realloc`
wrappers and the `calloc` wrapper can be used instead.