common pagecode: squash -Wswitch-default warning
`pagedir`, whose first character becomes the input to this switch, has legal
values “BL”, “BR”, “TL”, “TR”, “RB”, “RT”, “LB”, and “LT”. It can be set to any
other arbitrary string, but there is a follow-on warning in the caller
(`init_job_pagination`) if the returned point ends up as 0. This is the case
when we hit the default case here.

Gitlab: !3624
Suggested-by: Eric S. Raymond <esr@thyrsus.com>