cgraph agxbinit: remove support for null buffers
Callers who want an `agxbuf` with automatic storage can (and do) use zero
initialization (`agxbuf xb = {0}`). This is shorter and simpler.

The branch being removed contained some Undefined Behavior. MSVC noted:

  lib\cgraph\agxbuf.h(82): warning C6386: Buffer overrun while writing to
    'xb->store':  the writable size is '15' bytes, but '16' bytes might be
    written.

The overwrite here was intentional – it is setting the `located` member to
`AGXBUF_INLINE_SIZE_0` – but the way it was done is still technically UB. We
could fix this, but it seems simpler to remove this functionality.