arrows: fix slighty incorrect `lvee` and `rvee` edge arrow shapes for penwidths > 1
The `arrow_type_crow` function used one point too many when calling
the `gvrender_polygon` function.

This does not cause any problem for the `lcrow` and `rcrow` edge arrow
shapes since the `crow` "shaft" has the same visual width (i.e,
`penwidth`) as the edge "stem" (the three points (3, 4 & 5) defining
the end of the "shaft" are the same), while the `lvee` and `rvee`
arrow shapes for some reason don't.

The `vee` edge arrow shape is not affected since the surplus point (8)
is equal to the start point (0) and the `gvrender_polygon` function
always closes the polygon by adding the start point after the given
end point anyway.

This bug has existed since the first commit in the git history, but
before 2394c85e this made no difference since the "shaft" was then
defined with only one point, i.e., it had the visual width equal to
penwidth. Note that the issue number mentioned in that commit does not
seem to be the same issue as the corresponding GitLab issue. Commits
ef21f465 and 1e901e6f deals with the same issue and have somewhat more
meaningful commit messages.

Fixes https://gitlab.com/graphviz/graphviz/-/issues/2399.