fix: track matrix dimensions as 'size_t'
This resolves an integer overflow that occurs with large graphs. Such graphs are
typically still not (yet) processable, but may become so in future.

This required changing one instance of `flatindex` to `ND_low` because the
`flatindex` macro now uses a cast. We need to expand it out and remove the cast
in this one instance because it is used as an lvalue on the left side of an
assignment.

Gitlab: #1488
Reported-by: Róbert Papp