tests: svg_element: miter_shape: return full line join shape
This corrects the bounding box of the `crow` edge arrow shape, which
has sharp corners in the direction of the arrow. The miter length of
these would extend the miter limit and the line join shape therefore
falls back to bevel.

This causes a very small error to be detected in the
test_edge_node_overlap_normal_and_inv_edge_arrows_all_modifiers test
so it is temporarily set to be expectd to fail. This will be fixed in
the next commit in this series.