arrows: arrow_type_normal0: update fix of positioning of `normal`/`inv` arrow also for `l`/`r` modifiers
This is an improvement of the fix in
5a11aede5dc28e62ddc651ff01027e51efe58d49. It uses the real extreme
point of the miter shape which falls back to a bevelled corner for the
`l` and `r` modifiers.

This causes the
test_edge_node_overlap_normal_and_inv_edge_arrows_all_modifiers to
succeed again, so the expectancy that it should fail is removed.

Towards https://gitlab.com/graphviz/graphviz/-/issues/372.