add cmd, lib and dot.demo recursively to the INPUT
to avoid missing sources in the generated documentation
and simplify maintenance.

This commit adds many files to the generated documentation.

The EXCLUDE tag can be used to explicitly specify files
and/or directories that should be excluded from the INPUT source files.
This way you can easily exclude a subdirectory from a directory tree
whose root is specified with the INPUT tag.

I don't know whether exclusion of cmd/tools and other directories
was deliberate or not. I suppose it was by an accident.
I suppose the documentation should be inclusive by default
to avoid hiding information.

Plan is:
1. reveal everything
2. review
3. update brief descriptions of directories for clarification.
   (the briefs are shown in the listing of the parent directory)
4. EXCLUDE explicitly