gvgen: makeSquareGrid: simplify always-true and always-false code
The compiler notes that the `OUTE` expansion in the first branch here was
producing an always-false branch:

  graph_generator.c: In function 'makeSquareGrid':
  graph_generator.c:162:20: warning: assuming signed overflow does not occur
    when assuming that (X + c) < X is always false [-Wstrict-overflow]
   #define OUTE(h) if (tl < (hd=(h))) ef( tl, hd)
                    ^
  graph_generator.c:176:3: note: in expansion of macro 'OUTE'
     OUTE(i * dim2 + j);
     ^

Looking at the full body of this loop, the first four uses of `OUTE` all had
some variant of this situation. This change removes the always-false and
always-true branches, more explicitly spelling out the operations.

Gitlab: closes #2405