API BREAK: expand upper limit for minimum edge length to 'INT_MAX'
Users can reportedly run into the limit of `USHRT_MAX`¹ so this change relaxes
the limit. This change is based on 26ffd72fb80c3ff89e43d23ecfbbebd2f656147a. The
natural expansion would be to `unsigned`, but this value is used in calculations
with other `int` values so making it an `int` is simpler. `Agedgeinfo_t.minlen`
is never expected to be negative.

It is possible this has an effect on a legacy bug.
26ffd72fb80c3ff89e43d23ecfbbebd2f656147a claims to fix “bug 1034” but no
further information is given so it is unclear what this bug is/was. The bug
number is a reference to one of the past bug trackers that we no longer have
access to. There is no current issue referencing a bug with original ID 1034, so
the trail goes cold.

Gitlab: closes #2413

¹ https://forum.graphviz.org/t/edge-length-larger-than-maximum-allowed-on-linux-but-not-on-windows/1666