arrows: fix small gap between `lcurve`/`rcurve` shaft and node
Don't adjust the point of the shaft which is closest to the node based
on the penwidth. This avoids the small gap between the shaft and the
node on the side of the arrow where no curve is drawn. For the normal
`curve` this is not visible when the color is opaque since the
extension of the curve due to the penwidth covers this part.

Fixes https://gitlab.com/graphviz/graphviz/-/issues/2426.