CI: downgrade Python lint task to Ubuntu 22.04
Ubuntu 22.10 was EOLed on 2023-07-26.¹

The obvious thing would be to _upgrade_ to Ubuntu 23.04. But we run into the new
Python policy Ubuntu has instituted (see
0050eacd751dfc48bc3bc3c7af9f9c058a47e430). We could comply with this or move
this task to Fedora, but then we find that Pylint does not work with Python
3.11² which is what is in use on both Ubuntu 23.04 and all supported Fedora
releases. So downgrading to Ubuntu 22.04 seems like the simplest options for an
environment that is still in its support window.

¹ https://wiki.ubuntu.com/Releases
² Discussed in the comments of
  https://github.com/pylint-dev/pylint/issues/6535. Symptoms are complaints of
  `os` members not existing.