fix: gvc gvconfig: always initialize the font dictionary
Various paths through this function would result in leaving the font dictionary
uninitialized. When interacting with Graphviz libraries, a user program could
inadvertently trigger a lookup into this dictionary resulting in a confusing
crash.

This looks to have been an oversight in
a04ff22dd37c854be4576048e9ab17e8f57ad32d.

Note that the “rescan” code path is left as-is because there it is expected that
the font dictionary has already been initialized.

Gitlab: fixes #1520
Reported-by: Joseph Musser
Reported-by: Vithanco
Suggested-by: Cosmin Mihai