fix double escaping during canonicalization
Escape sequences that were present in strings like node labels were incorrectly
re-escaped during canonicalization. The effect of this was that, when using
Graphviz programmatically, there was no way to add labels containing certain
characters. For example `"`.

The change in this commit recognizes escape sequences and avoids re-escaping
them. This is not ideal, but seems the best we can do with ambiguous input (it
is not strictly required for the input to be either escaped or unescaped).
Contradicting the official documentation,¹ this change recognizes `\"` as an
escape sequence for `"`. Without this, I see no way to represent `"` within a
label.

This test still needs to be xfailed on Windows because for some reason the test
binary cannot be compiled and linked.

Gitlab: #165, fixes #2397
Reported-by: Nikolai Kondrashov

¹ In particular, this sentence from
  https://graphviz.org/docs/attr-types/escString/, “A backslash appearing before
  any character not listed above is ignored.”