ingraphs: remove unnecessary hacks to avoid including cgraph.h
Apart from being repetitive and error prone because definitions could get out of
sync, the introduction of a different `Agraph_t` type and then using it as an
alias to the real `Agraph_t` is Undefined Behavior. It is unlikely this would
have caused problems as the two definitions are in separate translation units,
so only an LTO compilation would be able to see this discrepancy. But it is
safer to remove this.