API BREAK: remove memory allocator disciplines
A memory allocator discipline is a per-graph set of functions describing how to
allocate and deallocate heap memory for a graph. It allows specifying a custom
allocator for each graph.

When using this mechanism to substitute a custom allocator to replace malloc, it
was insufficient. Graphviz directly calls `malloc`/`calloc`/etc in numerous
places, ignoring memory allocator discipline choices.

In theory, memory allocator disciplines can be used to bootstrap Graphviz on a
platform without dynamic memory allocation. However doing this requires also
overriding `malloc` and friends because, as discussed above, Graphviz directly
calls these functions in numerous places. Globally swapping the system allocator
requires the use of a mechanism like `LD_PRELOAD`. This is true both before and
after the current change.

No Graphviz code uses this mechanism. Everything uses the `AgMemDisc` discipline
which backs all allocation with `calloc`.

The only uses of the symbols removed in this commit outside Graphviz itself
found by Debian Code Search¹ are in PlantUML’s automatically translated Graphviz
headers. This is a strong indication that users will be minimally affected by
this change if at all.

This change also fixes a number of unchecked allocation failures by swapping a
(indirect) `calloc` for `gv_calloc`.

¹ https://codesearch.debian.net/