CMake: remove unused 'LTDL_H'
Introduced in 3b567c4bd938e2e70478b22d5efae98bedf9ed8f, this seems to have been
attempting to create a `#define` that the code sees. But `add_definitions` is
the way to do this, not setting a CMake variable. Moreover, it is incorrect to
try to define this here anyway. The one use of `LTDL_H` in the code is
attempting to check for a macro defined in the ltdl.h header, not in Graphviz
configuration.