ortho decide_point: fix uninitialized read of 'np2'
MSVC says:

  C:\GitLab-Runner\builds\graphviz\graphviz\lib\ortho\ortho.c(849): warning
    C4701: potentially uninitialized local variable 'np2' used
    [C:\GitLab-Runner\builds\graphviz\graphviz\lib\ortho\gvortho.vcxproj]

It is not clear if this is a false positive or not. That is, whether the initial
execution of the preceding `while` loop can encounter `np1 == NULL` and never
perform the assignment to `np2`. The content of the body of the `if` here
suggests maybe this code is unfinished or expected to be unreachable.