sfio setclass: fix: read input as unsigned chars
Characters from the input are used to index into the array `accepted`. However
it is implementation defined whether `char` is signed or unsigned. On platforms
with a signed character, these indexing operations can read out-of-bounds at a
negative offset before the start of `accepted`.

This code originates from a time before UTF-8. So negative character values (the
high bit set) were not generally expected. Nowadays bytes like this are common
in text data.