API BREAK: cdt: unravel and remove 'memoryf'
This always bottoms out on `malloc` and friends. Debian Code Search¹ finds only
uses of this in Graphviz itself, GCC test cases presumably derived from this
code, ksh which contains its own copy of libast, and Plantuml comments. This is
a strong indication no one is using this. Removing it simplifies code and eases
ongoing maintenance.

¹ https://codesearch.debian.net/