ast pathaccess: only canonicalize path after existence check
The `access` and `stat` calls do not need canonicalized paths. This is slightly
more efficient (not canonicalizing paths we never need), but the real motivation
is to move this operation to somewhere that we know the path exists and thus can
replace this with a call to `realpath` in an upcoming commit.