ast pathaccess: replace 'pathcanon' call with `realpath'/`_fullpath`
The main downside of `realpath` is that it does not work on paths that do not
point to existing files. However in this branch we already know that the path
points to an existing file. This change removes the only use of `pathcanon`,
allowing it to be removed in an upcoming commit.

This change removes a limitation of this code that prevented resolution to a
target path greater than `PATH_MAX` bytes. glibc’s `realpath` man page
summarizes some of the problems with trying to resolve a path into a `PATH_MAX`
sized buffer:

  The POSIX.1-2001 standard version of this function is broken by design, since
  it is impossible to determine a suitable size for the output buffer,
  `resolved_path`. According to POSIX.1-2001 a buffer of size `PATH_MAX`
  suffices, but `PATH_MAX` need not be a defined constant, and may have to be
  obtained using `pathconf(3)`. And asking `pathconf(3)` does not really help,
  since, on the one hand POSIX warns that the result of `pathconf(3)` may be
  huge and unsuitable for `malloc`ing memory, and on the other hand
  `pathconf(3)` may return -1 to signify that `PATH_MAX` is not bounded. The
  `resolved_path == NULL` feature, not standardized in POSIX.1-2001, but
  standardized in POSIX.1-2008, allows this design problem to be avoided.

Gitlab: #2452