cgraph gv_sort: support null 'base' pointer
Calling `qsort(NULL, …)` is undefined behavior. Surprisingly, even if `nmemb`
is 0, a situation in which it is natural to end up passing `NULL` as `base`.
Though it seems only possible to realize this by implication of the wording of
ISO C99 §7.20.5.2.2:

  The `qsort` function sorts an array of `nmemb` objects, the initial element of
  which is pointed to by `base`.

I myself only learned this from the Undefined Behavior Sanitizer (UBSan), that
helpfully notices `qsort(NULL, 0, …)` calls are a problem due to `qsort`’s first
parameter being tagged non-null.

By avoiding an invocation of `qsort` in this scenario, we can make
`gv_sort(NULL, 0, …)` well-defined, allowing more natural caller code.

The added conditional is `nmemb > 1` instead of the perhaps expected `nmemb > 0`
because sorting ≤ 1 members is a no-op.