tclpkg: fix package version numbering to use 'b' instead of '~dev.' as separator
Graphviz releases use semantic versioning, e.g. 8.2.4. Inter-release builds use
the next upcoming release’s version with an appended developer suffix, e.g.
8.2.4~dev.20231012.1016.¹ The presence of the characters `~dev` caused TCL to
reject loading the Graphviz packages as it considers these part of an invalid
version string. The effect of this was that Graphviz TCL packages were only
usable in official releases, not in builds compiled from arbitrary Git commits.

Resolving this is surprisingly complicated. A non-exhaustive list of
requirements:

  1. TCL versions are only permitted to contain numbers, `.`, and the letters
    `a` and `b`.²
  2. The RPM and DEB ecosystems sort packages pseudo-alphabetically, with `~`
     and `rc` counting as earlier than an end of string.³ We want our
     inter-release versions to sort _before_ their upcoming official release.
  3. The RPM ecosystem forbids `-` in version numbers and the DEB ecosystem
     forbids everything except alphanumerics, `.`, `+`, `-`, `:`, and `~`.⁴

If you put just (1) and (2) together, you will realize there is no way to
construct an inter-release version string that begins with the upcoming release
version, but is valid in all these packaging ecosystems and also sorts before
the upcoming release.

This change chooses to use a differing versioning scheme for the Graphviz TCL
packages. `~dev.` is replaced with `b`. Taking the example above, version
8.2.4~dev.20231012.1016 would be reported by the Graphviz TCL packages as
8.2.4b20231012.1016. Though these strings are clearly not the same, I believe
they give us the desired properties:

  1. 8.2.4~dev.20231012.1016 is valid in the Linux ecosystem
  2. The Linux ecosystems sort 8.2.4~dev.20231012.1016 before 8.2.4
  3. 8.2.4b20231012.1016 is valid in TCL
  4. TCL sorts 8.2.4b20231012.1016 before 8.2.4

A possible lens in which to view these sorting schemes is the way TCL describes
them,² with some components contributing negative values:

  the letters “a” (alpha) and/or “b” (beta) may appear exactly once to replace a
  dot for separation. These letters semantically add a negative specifier into
  the version, where “a” is -2, and “b” is -1.

A reasonable question to ask is why any of the Linux packaging concerns are
relevant, given this is about inter-release versions. While it is true that
today (AFAIK) no Linux distribution packages non-release versions of Graphviz,
this may not be true in future. It would be beneficial to give downstream
maintainers as many degrees of freedom as possible and allow them to package
inter-release versions if they choose. The more boring answer is that Graphviz
CI builds Linux packages for every commit and the tooling that does this needs
to see a compliant version. It is also possible that a non-release version is
accidentally packaged for Linux. E.g. Debian/Ubuntu offers a Graphviz package
listed as 2.42.2, but package itself claims to be 2.43.0.

Gitlab: fixes #2370
Reported-by: Dario Gjorgjevski

¹ See gen_version.py for the full details.
² https://www.tcl.tk/man/tcl/TclCmd/package.html#M20
³ https://github.com/semver/semver/issues/145#issuecomment-35662854
  Github has turned `~ … ~` into a request for strike-through formatting, so you
  will have to think a bit to realize what the commenter actually wrote.
⁴ https://github.com/semver/semver/issues/145#issuecomment-109044744