remove parentheses that may inhibit NRVO
The idiom `return (foo)`, that is wrapping a return expression in parentheses,
is a C technique to let you define `return` as a macro to alter its behavior.
This has mostly fallen out of favor in contemporary times as it is considered
confusing and there are better ways to achieve this kind of return
instrumentation/redirection.

Named Return Value Optimization (NRVO)¹ is a C++ feature that allows
constructing a value-to-be-returned directly in its final target location in
certain circumstances. The scenario relevant to this commit is discussed in ISO
C++11 §12.8.31:

  In a `return` statement in a function with a class return type, when the
  expression is the name of a non-volatile automatic object (other than a
  function or catch-clause parameter) with the same cv-unqualiﬁed type as the
  function return type, the copy/move operation can be omitted by constructing
  the automatic object directly into the function’s return value

Surprisingly the seemingly benign parentheses in `return (foo)` interfere with
this optimization. It seems unclear whether they _should_ and even compilers
disagree on this.² The code touched in this commit is not returning non-POD
values or relying on NRVO, but removing the parentheses to avoid this ambiguous
situation altogether seems wise.

This commit only touches C++ code that exhibits this pattern, but we should
probably remove this idiom from C code as well for the sake of readability.

¹ https://en.cppreference.com/w/cpp/language/copy_elision
² https://stackoverflow.com/questions/48749440/c-nrvo-copy-elision-with-return-statement-in-parentheses