cmake: make 'with_zlib' tristate
Instead of being `ON` or `OFF`, `with_zlib` can now also have the additional
value `AUTO`. The new value `AUTO` has the semantics of what was previously `ON`
(enable zlib stuff if zlib is found). The new behavior of `ON` is to fail the
build if zlib is not found. `OFF` avoids looking for zlib at all. This gives
users more control and fewer surprises when building Graphviz.

Gitlab: #2312