cmake: make 'enable_ltdl' tristate
Instead of being `ON` or `OFF`, `enable_ltdl` can now also have the additional
value `AUTO`. The new value `AUTO` has the semantics of what was previously `ON`
(enable ltdl stuff if ltdl is found). The new behavior of `ON` is to fail the
build if ltdl is not found. `OFF` avoids looking for ltdl at all. This gives
users more control and fewer surprises when building Graphviz.

Gitlab: #2312