cmake: make 'with_expat' tristate
Instead of being `ON` or `OFF`, `with_expat` can now also have the additional
value `AUTO`. The new value `AUTO` has the semantics of what was previously `ON`
(enable expat stuff if expat is found). The new behavior of `ON` is to fail the
build if expat is not found. `OFF` avoids looking for expat at all. This gives
users more control and fewer surprises when building Graphviz.

Gitlab: #2312