CI: set '-Denable_ltdl=ON' in most CMake builds
The only one omitted is the minimal build which already has `-Denable_ltdl=OFF`.
Setting `ON` explicitly instead of the previous `AUTO` behavior ensures CI fails
if ltdl is accidentally removed or unavailable.

An alternative to the changes to .gitlab-ci.yml would have been adding this to
the CMake invocation in ci/build.sh. But it seems less confusing to keep these
tweaks in one (well, four) place.

Gitlab: #2312