GTK plugin: remove build system support
This plugin depends on GTK 2 that is deprecated and disappearing from ecosystems
and X11 that is being replaced by Wayland. It is suspected it has few or no
users. This change disables it as a gradual step towards removing it or learning
of active users.

This follows a similar staged approach to how Lefty was removed:

  1. Default flipped to disabled (1a4ad00b59cc1e1e9302fcbd1a205495ee59cab5)
  2. Build system support removed (6faa34010f947510ec129af2d998614de8a513ed)
  3. Code and final pieces removed (b2f287ca315dd5bf689d9ce2ca71dbfe63fb8a61)

but jumps straight to (2). There is no configuration flag for the GTK plugin, so
no way to do (1). This change is less invasive than
6faa34010f947510ec129af2d998614de8a513ed, in that it just disconnects the GTK
plugin build system components, but does not remove them. If we learn of
downstream users, we can revert this and work on forward porting the plugin.

Even if this plugin is removed, it will not fully remove Graphviz’ GTK
dependency because Smyrna also depends on GTK.

Gitlab: #1484