Autotools: fix pkg-config-less Ruby discovery
This effectively reverts part of 3a85bbcbcdbbb0968a806320e64b40c0fe3feb80 which
seems to be incorrect. `includedir`, that it removed, was still not correct but
was closer than `archdir` which hosts libraries.

This commit was arrived at by probing the `MAKEFILE_CONFIG` dictionary for what
relevant looking variables exist and the order in which they occur. It was
tested by removing the pkg-config logic from configure.ac and confirming the
build system can find a default Ruby 3.0.0 installation on Ubuntu 22.04.

Gitlab: fixes #2464
Reported-by: Zachary Gillions