Autotools: support suppressing 'LIBPOSTFIX'
The Autotools build system allows `LIBPOSTFIX` to be set to indicate a suffix to
append to the installation library path. If it is unset, a default of `64` or
`/64` is used on various platforms. However the logic for this did not
discriminate between `LIBPOSTFIX` being unset vs it being set to the empty
string. Thus on a platform where one of these defaults applied, there was no way
to indicate you wanted no suffix.

This is relevant to platforms like Linux From Scratch (LFS)¹ that use a unified
library directory:²

  The LFS editors have deliberately decided not to use a `/usr/lib64` directory.
  Several steps are taken to be sure the toolchain will not use it. If for any
  reason this directory appears (either because you made an error in following
  the instructions, or because you installed a binary package that created it
  after finishing LFS), it may break your system. You should always be sure this
  directory does not exist.

This is perfectly acceptable under the File System Hierarchy (FHS) standard,³
but the Graphviz build system did not support this adequately. As a consequence,
guides like Beyond Linux From Scratch are advising manually patching the
Graphviz build system to work around this:⁴

  Install Graphviz by running the following commands:

    ```
    sed -i '/LIBPOSTFIX="64"/s/64//' configure.ac &&
    ./autogen.sh &&
    ./configure --prefix=/usr \
       --docdir=/usr/share/doc/graphviz-8.1.0
    ```

  Note
  A warning is generated by `autogen.sh` because the build tree is not a git
  repository. As a result, the build date is set to 0. To get a meaningful date
  in the version string, you can run:

    ```
    sed -i "s/0/$(date +%Y%m%d)/" builddate.h
    ```

  …

  `sed ... configure.ac`: This command is needed to avoid installing files in
  /usr/lib64.

This change gives users a supported way of suppressing this suffix,
`./configure LIBPOSTFIX=`.

Gitlab: #2464

¹ https://www.linuxfromscratch.org/
² https://www.linuxfromscratch.org/lfs/view/development/chapter04/creatingminlayout.html
³ https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch03s10.html
⁴ https://lfs.maru-na.net/blfs/downloads/stable-systemd/BLFS-BOOK-12.0-systemd-nochunks.html#graphviz