fix truncation of xdot fields containing escaping
Xdot strings are encoded in a format that has their length preceding them. This
lets readers of such fields (including Graphviz itself) pre-allocate memory for
the upcoming string. Unfortunately the length is given in units of unescaped
(original) characters and the string itself is escaped. E.g. a string “f\oo”
would be encoded as “f\\oo”, but its length given as 4. There were two primary
negative side effects of this:

  1. The pre-allocation for the upcoming string was insufficient in these
     scenarios.

  2. The number of bytes parsed was truncated. In the previous example, only
     “f\\o” would be read.

This problem emerged to users in surprising ways because Graphviz encodes xdot
fields and then parses them back, even when not generating xdot. The report that
led to this discovery was against the JSON output format, something one would
have guessed had nothing to do with xdot.

This fix treats the preceding length of an xdot string as a lower bound rather
than a precise count of the upcoming characters.

A curious reader may have wondered why the above scenario does not lead to a
trailing “o” that then confuses parsing the next field. This seems to be avoided
because xdot fields are being parsed one-by-one in this situation (JSON
rendering). The trailing “o” is ignored because the parser thinks it is done
with the current unit.

Gitlab: fixes #2460
Reported-by: Adrian Bird