CMake: fix macOS builds by using '*_LINK_LIBRARIES' in rsvg plugin
The `FOO_LIBRARIES` variables contain strings like `-lfoo -lbar`, while the
`FOO_LINK_LIBRARIES` variables contain strings like
`-l/path/to/foo.dylib -l/path/to/bar.dylib`. On macOS, the library paths the
compiler searches do not seem to be setup to include the paths that would
contain foo.dylib and bar.dylib. This commit alters the CMake build to use
absolute paths, so the libraries are correctly found.

(It is not clear to me this is the right fix. This repairs macOS builds, but
also prompts the question of why FindPANGOCAIRO.cmake/FindRSVG.cmake can find
these libraries if the compiler subsequently cannot. Discussion in the issues
linked below suggests there may be something deeper going on.)

Gitlab: #2286, #2290, #2411