Fix UBSAN
vec[c].size can be 0 in some cases, even if vec[c].list != NULL. The
loop does not execute if the size is 0, because i will be initialized
to 0 - 1 as size_t hence SIZE_MAX, and the intention is to skip
the loop as now i != SIZE_MAX check fails. However since the loop
initializers run first, we will try to do vec[c].list + i, which add
SIZE_MAX to whatever list points, to and it will overflow.

The fix here is to add a size check. I think we can remove the
vec[c].list check, as it should be impossible that we have
vec[c].list != NULL && vec[c].size == 0.