common: reset node and edge lists when existing network simplex
This leads to slightly easier to reason about code, with scratch memory here not
reused across calls into this code.

Seeing why this change works requires tracing paths through `rank2`. Every one
either directly or indirectly leads to a call to `freeTreeList` except the first
case in the switch statement. That is why it is the only one to have a
`reset_lists` call inserted.