remove support for OCaml bindings
SWIG added support for OCaml in version 1.3.14¹ released 2002-08-12.² This
support has been marked “Experimental” and still is so as of the end of 2023.³

Graphviz gained similarly experimental OCaml bindings around
753b4dbddf3501169ea4c0351c556b88ecb7395d in 2005. At present, there seems no
evidence of users of these bindings, yet they are a recurring source of build
problems.⁴

This commit removes the generation of these bindings from the Autotools build
system (they are not present in the CMake or MS Build build systems) and
uninstalls their dependencies from CI images. If users make their presence known
in response to this change, we can revert this.

Gitlab: closes #2430

¹ https://www.swig.org/compat.html
² https://www.swig.org/history.html
³ See `-ocaml` in Source/Modules/swigmain.cxx in https://github.com/swig/swig at
  commit ee4f03117760f8d31c25e74278cff0ddc8e2ef65.
⁴ #641, #1027, #1037, #1262, #2128, #2234, #2430